Rename acms-orders use case to migrate-consolidation
Jira ticket: CAMS-421
