+++ backend/function-apps/dataflows/migration/activity/flattenBoundingArrays.test.ts
@@ -1,5 +1,8 @@
 import { InvocationContext } from '@azure/functions';
-import { AcmsBounds, AcmsPredicate } from '../../../../lib/use-cases/acms-orders/acms-orders';
+import {
+  AcmsBounds,
+  AcmsPredicate,
+} from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
 import module from './flattenBoundingArrays';
 
 describe('Flatten bounding arrays activity', () => {
