+++ backend/function-apps/dataflows/migration/activity/getConsolidations.test.ts
@@ -1,6 +1,6 @@
 import AcmsOrdersController from '../../../../lib/controllers/acms-orders/acms-orders.controller';
 import queueMigrateConsolidation from './queueMigrateConsolidation';
-import { AcmsPredicate } from '../../../../lib/use-cases/acms-orders/acms-orders';
+import { AcmsPredicate } from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
 import { createMockAzureFunctionContext } from '../../../azure/testing-helpers';
 import { CamsError } from '../../../../lib/common-errors/cams-error';
