+++ backend/function-apps/dataflows/migration/orchestration/orchestrator.ts
@@ -1,4 +1,7 @@
-import { AcmsBounds, AcmsPredicate } from '../../../../lib/use-cases/acms-orders/acms-orders';
+import {
+  AcmsBounds,
+  AcmsPredicate,
+} from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
 import { FLATTEN_BOUNDING_ARRAYS, SUB_ORCHESTRATOR_ETL } from '../migration';
 import { OrchestrationContext } from 'durable-functions';
