+++ backend/function-apps/dataflows/migration/queueTrigger/migrateConsolidations.test.ts
@@ -5,7 +5,7 @@ import { CamsError } from '../../../../lib/common-errors/cams-error';
 import {
   AcmsEtlQueueItem,
   AcmsTransformationResult,
-} from '../../../../lib/use-cases/acms-orders/acms-orders';
+} from '../../../../lib/use-cases/acms-orders/migrate-consolidations';
 
 describe('getConsolidations test', () => {
   afterEach(() => {
